added loop for confirming choice
Had a problem where the suggestion was getting turned down.  If this happened, an entry was needlessly put into the atethere.txt list.  Added a loop to add a vote to confirm decision.
Input of Yay, y, or Y confirms suggestion as "we are going eat there, put it in the list"